Stage 1: Separated master/detail from main view into their own view
